+++ src/muleengine/mesh_creator/capsule.rs
@@ -83,14 +83,14 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
         );
 
         mesh.add_face(
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 3,
-            mesh.number_of_vertices() - 4,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 3) as u32,
+            (mesh.number_of_vertices() - 4) as u32,
         );
         mesh.add_face(
-            mesh.number_of_vertices() - 1,
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 4,
+            (mesh.number_of_vertices() - 1) as u32,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 4) as u32,
         );
     }
 
@@ -179,14 +179,14 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
             );
 
             mesh.add_face(
-                mesh.number_of_vertices() - 2,
-                mesh.number_of_vertices() - 3,
-                mesh.number_of_vertices() - 4,
+                (mesh.number_of_vertices() - 2) as u32,
+                (mesh.number_of_vertices() - 3) as u32,
+                (mesh.number_of_vertices() - 4) as u32,
             );
             mesh.add_face(
-                mesh.number_of_vertices() - 1,
-                mesh.number_of_vertices() - 2,
-                mesh.number_of_vertices() - 4,
+                (mesh.number_of_vertices() - 1) as u32,
+                (mesh.number_of_vertices() - 2) as u32,
+                (mesh.number_of_vertices() - 4) as u32,
             );
         }
     }
@@ -251,9 +251,9 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
         );
 
         mesh.add_face(
-            mesh.number_of_vertices() - 1,
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 3,
+            (mesh.number_of_vertices() - 1) as u32,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 3) as u32,
         );
     }
 
@@ -338,14 +338,14 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
             );
 
             mesh.add_face(
-                mesh.number_of_vertices() - 4,
-                mesh.number_of_vertices() - 3,
-                mesh.number_of_vertices() - 2,
+                (mesh.number_of_vertices() - 4) as u32,
+                (mesh.number_of_vertices() - 3) as u32,
+                (mesh.number_of_vertices() - 2) as u32,
             );
             mesh.add_face(
-                mesh.number_of_vertices() - 4,
-                mesh.number_of_vertices() - 2,
-                mesh.number_of_vertices() - 1,
+                (mesh.number_of_vertices() - 4) as u32,
+                (mesh.number_of_vertices() - 2) as u32,
+                (mesh.number_of_vertices() - 1) as u32,
             );
         }
     }
@@ -408,9 +408,9 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
         );
 
         mesh.add_face(
-            mesh.number_of_vertices() - 3,
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 1,
+            (mesh.number_of_vertices() - 3) as u32,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 1) as u32,
         );
     }
