+++ src/muleengine/mesh_creator/cylinder.rs
@@ -80,14 +80,14 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
         );
 
         mesh.add_face(
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 3,
-            mesh.number_of_vertices() - 4,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 3) as u32,
+            (mesh.number_of_vertices() - 4) as u32,
         );
         mesh.add_face(
-            mesh.number_of_vertices() - 1,
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 4,
+            (mesh.number_of_vertices() - 1) as u32,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 4) as u32,
         );
     }
 
@@ -142,9 +142,9 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
         );
 
         mesh.add_face(
-            mesh.number_of_vertices() - 1,
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 3,
+            (mesh.number_of_vertices() - 1) as u32,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 3) as u32,
         );
     }
 
@@ -199,9 +199,9 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
         );
 
         mesh.add_face(
-            mesh.number_of_vertices() - 1,
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 3,
+            (mesh.number_of_vertices() - 1) as u32,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 3) as u32,
         );
     }
